Make all mouse buttons and scrolling count as "activity" (#1)
Currently, the `click` event fires for all mouse buttons, but "soon" it'll fire only for the primary mouse button. `mousedown` fires for all buttons and will continue to do so (in the foreseeable future).

Scrolling seems like it should count as activity, so now it does! The docs note that browsers can generate a lot of scroll events in a short time and suggest throttling for "expensive" computations, but what we're doing with them seems cheap enough.